fix: imports after moving components
