+++ src/components/post/types.ts
@@ -1,5 +1,5 @@
 import type { Post as PrismaPost } from "@prisma/client"
-import { type PostAuthor } from "../profilePage/types"
+import { type PostAuthor } from "../../features/profile/types"
 import { type RouterOutputs } from "~/utils/api"
 
 export type PostWithAuthor = {
