+++ src/server/features/homePage/fetch.ts
@@ -1,7 +1,7 @@
 import { TRPCError } from "@trpc/server"
 import type { PostWithAuthor } from "~/components/post/types"
-import { type PostAuthor } from "~/components/profilePage/types"
 import { CONFIG } from "~/config"
+import { type PostAuthor } from "~/features/profile"
 import { getUserFollowList } from "~/server/api/follow"
 import { getPostById } from "~/server/api/posts"
 import {
